Bunch of crap related to, or not, to auto service list

Approved-by: Brad Richardson

+++ app/src/main/java/com/carswaddle/carswaddleandroid/services/serviceModels/Point.kt
@@ -2,14 +2,14 @@ package com.carswaddle.carswaddleandroid.services.serviceModels
 
 data class Point (
     val type: String,
-    val coordinates: Array<Float>
+    val coordinates: Array<Double>
 ) {
 
-    fun latitude(): Float {
+    fun latitude(): Double {
         return coordinates[0]
     }
 
-    fun longitude(): Float {
+    fun longitude(): Double {
         return coordinates[1]
     }
